+++ app/(home)/user/settings/edit-image.tsx
@@ -1,0 +1,5 @@
+import { EditImageScreen } from 'appdeptus/modules/user/Screens'
+
+const EditImageRoute = () => <EditImageScreen />
+
+export default EditImageRoute