+++ modules/user/Screens/index.ts
@@ -1,5 +1,6 @@
 export { default as AdeptScreen } from './Adept'
 export { default as DeleteScreen } from './Delete'
+export { default as EditImageScreen } from './EditImage'
 export { default as EditNameScreen } from './EditName'
 export { default as SettingsScreen } from './Settings'
 export { default as UserScreen } from './User'