+++ modules/user/api/endpoints/index.ts
@@ -1,2 +1,3 @@
 export { default as getUserProfile } from './getUserProfile'
+export { default as updateUserImage } from './updateUserImage'
 export { default as updateUserName } from './updateUserName'