+++ modules/user/api/schemas.ts
@@ -4,7 +4,8 @@ const userProfileSchema = z
   .object({
     id: z.string(),
     name: z.string(),
-    created_at: z.string()
+    created_at: z.string(),
+    image: z.string().optional()
   })
   .transform(({ created_at, ...rest }) => ({ ...rest, createdAt: created_at }))