+++ utils/index.ts
@@ -1,3 +1,4 @@
 export { default as mapNullToUndefined } from './mapNullToUndefined'
 export { default as mapRosterToUnits } from './mapRosterToUnits'
+export { default as selectImageFromGallery } from './selectImageFromGallery'
 export { default as supabase } from './supabase'